requirements.txt
anyio==3.6.2
async-timeout==4.0.2
beautifulsoup4==4.12.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
fastapi==0.95.1
h11==0.14.0
idna==3.4
jaconv==0.3.4
markdown-it-py==2.2.0
mdurl==0.1.2
mutagen==1.46.0
platformdirs==3.5.0
pycryptodomex==3.17
pydantic==1.10.7
Pygments==2.15.1
pykakasi==2.2.1
python-dotenv==1.0.0
python-slugify==8.0.1
pytube==12.1.3
rapidfuzz==2.15.1
redis==4.5.4
requests==2.29.0
rich==13.3.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
spotdl==4.1.8
spotipy==2.23.0
starlette==0.26.1
syncedlyrics==0.5.0
text-unidecode==1.3
typing_extensions==4.5.0
Unidecode==1.3.6
urllib3==1.26.15
uvicorn==0.21.1
websockets==11.0.2
wrapt==1.15.0
yt-dlp==2023.3.4
ytmusicapi==0.24.1
